Visual Bug Fix
fixed visual bug that caused players money to display wrong value after gaining their winnings

added new method to Player.cs to handle all money updates instead of Dealer directly accessing money values on players. This is also used as a replacement to GiveWinnerEarnings as it was not working correctly.